test: merge cors tests
